% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cochrane_orcutt.R
\name{cochrane_orcutt}
\alias{cochrane_orcutt}
\title{Title of the function}
\usage{
cochrane_orcutt(reg, convergence = 8, max.iter = 100)
}
\arguments{
\item{x}{Description of parameter x.}
}
\value{
What the function returns.
}
\description{
Description of what the function does.
}
