% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.orcutt.R
\name{predict.orcutt}
\alias{predict.orcutt}
\title{Predict method for Cochrane-Orcutt estimation}
\usage{
\method{predict}{orcutt}(object, ...)
}
\arguments{
\item{object}{An object of class \code{orcutt}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A numeric vector of fitted values.
}
\description{
Returns the fitted values from a Cochrane-Orcutt model object.
}
