% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_orcutt.R
\name{predict_orcutt}
\alias{predict_orcutt}
\title{Title}
\usage{
predict_orcutt(object, ...)
}
\arguments{
\item{object}{An object of class \code{orcutt}.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A numeric vector of fitted values.
}
\description{
Title
}
\examples{
model <- cochrane.orcutt(lm(y ~ x, data = icecream))
predict(model)
}
