% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.orcutt.R
\name{print.orcutt}
\alias{print.orcutt}
\title{Print method for Cochrane-Orcutt estimation}
\usage{
\method{print}{orcutt}(x, ...)
}
\arguments{
\item{x}{An object of class \code{orcutt}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Displays results from a Cochrane-Orcutt model fit.
}
